requirements.txt
openpyxl==3.1.2
python-docx==1.1.0
# tkinter - já incluído no Python